requirements.txt
numpy
torch
torchvision
transformers
Flask
Pillow
requests
matplotlib
notebook
